% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{pie.effects}
\alias{pie.effects}
\title{pie.effects}
\usage{
pie.effects(pier, ...)
}
\arguments{
\item{pier}{object}

\item{load}{list with options:
\itemize{
 \item{effect}{: character one of 'none', 'default}
 \item{speed}{: numeric milliseconds}
}}

\item{pullOutSegmentOnClick}{list with options:
\itemize{
 \item{effect}{: character one of 'none', 'linear', 'bounce', 'elastic', 'back'}
 \item{speed}{: integer milliseconds}
 \item{size}{: integer pixels}
}}

\item{highlightSegmentOnMouseover}{boolean}

\item{highlightLuminosity}{numeric}
}
\description{
Pass effects options to pier object
}
\examples{
\dontrun{
data.frame(value = c(40, 20, 30),
           label = c('red', 'green', 'blue'),
           color = c('red', 'green', 'blue')) \%>\%
  pier() \%>\%
  pie.effects(load  = list(speed = 2000),
              highlightLuminosity = .2)
  }
}

