% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{pie.footer}
\alias{pie.footer}
\title{pie.footer}
\usage{
pie.footer(pier, text = NULL, size = NULL, font = NULL, colour = NULL,
  location = NULL)
}
\arguments{
\item{pier}{object}

\item{text}{character main title text}

\item{size}{integer font size}

\item{font}{character font name; eg. "Impact"}

\item{colour}{character colour name or hex value}

\item{location}{character; must be one of 'bottom-left','bottom-center','bottom-right'}
}
\description{
Add footer to pie
}
\examples{
\dontrun{
data.frame(value = c(40, 20, 30),
           label = c('red', 'green', 'blue'),
           color = c('red', 'green', 'blue')) \%>\%
  pier() \%>\%
  pie.footer('with some additional explanation')
  }
}
