% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{pie.header}
\alias{pie.header}
\title{pie.header}
\usage{
pie.header(pier, text = NULL, size = NULL, font = NULL, colour = NULL,
  location = NULL)
}
\arguments{
\item{pier}{object}

\item{text}{character main title text}

\item{size}{integer font size}

\item{font}{character font name; eg. "Impact"}

\item{colour}{character colour name or hex value}

\item{location}{character; must be one of 'top-center', 'top-left' or 'pie-center'}
}
\examples{
\dontrun{
data.frame(value = c(40, 20, 30),
           label = c('red', 'green', 'blue'),
           color = c('red', 'green', 'blue')) \%>\%
  pier() \%>\%
  pie.header(text = 'My New Title', font = 'Impact',
             size = 18, colour = 'grey',
             location = 'top-center')
  }
}
