% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{pie.labels}
\alias{pie.labels}
\title{pie.labels}
\usage{
pie.labels(pier, ...)
}
\arguments{
\item{pier}{object}

\item{outer}{list with options:
\itemize{
 \item{format}{character default 'label'}
 \item{hideWhenLessThanPercentage}{: boolean default NULL}
 \item{pieDistance}{: numeric pixels}
}}

\item{inner}{list with options:
\itemize{
 \item{format}{: character default 'percentage'}
 \item{hideWhenLessThanPercentage}{: boolean default NULL}
}}

\item{mainLabel}{list with options:
\itemize{
 \item{color}{: character default '#333333'}
 \item{font}{: character default 'arial'}
 \item{fontSize}{: integer pixels}
}}

\item{percentage}{list with options:
\itemize{
 \item{color}{: character default '#dddddd'}
 \item{font}{: character default 'arial'}
 \item{fontSize}{: integer pixels}
 \item{decimalPlaces}{: integer}
}}

\item{value}{list with options:
\itemize{
 \item{color}{: character default '#cccc44'}
 \item{font}{: character default 'arial'}
 \item{fontSize}{: integer pixels}
}}

\item{lines}{list with options:
\itemize{
 \item{enabled}{: boolean}
 \item{style}{: character default }
 \item{color}{: character "segment" or a hex color}
}}
}
\description{
Pass Labelling options to pier object
}
\examples{
\dontrun{
data.frame(value = c(40, 20, 30),
           label = c('red', 'green', 'blue'),
           color = c('red', 'green', 'blue')) \%>\%
  pier() \%>\%
  pie.labels(mainLabel = list(font = 'Impact',
                              size = 14))
  }
}

