% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{pie.misc}
\alias{pie.misc}
\title{Miscellaneous Settings}
\usage{
pie.misc(pier, ...)
}
\arguments{
\item{pier}{object}

\item{...}{variables all passed via ...}

\item{colors}{list with options:
\itemize{
 \item{background}{: character color name or hex. Default is NULL results in tansparent}
 \item{segments}{: character vector segment colors only works where not defined in data}
 \item{segmentStroke}{: character color name or hex}
}}

\item{gradient}{list with options:
\itemize{
 \item{enabled}{: boolean}
 \item{percentage}{: integer gradient percent}
 \item{color}{: character color name or hex}
}}

\item{canvasPadding}{list with options:
\itemize{
 \item{top}{: integer pixels}
 \item{right}{: integer pixels}
 \item{bottom}{: integer pixels}
 \item{left}{: integer pixels}
}}

\item{pieCenterOffset}{list with options:
\itemize{
 \item{x}{: integer pixels}
 \item{y}{: integer pixels}
}}

\item{cssPrefix}{NULL}
}
\description{
Pass miscellaneous options to pier object
}
\examples{
data.frame(value = c(40, 20, 30),
           label = c('red', 'green', 'blue'),
           color = c('red', 'green', 'blue')) \%>\%
  pier() \%>\%
  pie.misc(gradient  = list(enabled = TRUE))

}
