% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie.R
\name{pie}
\alias{pie}
\title{Create elegant interactive pie charts.}
\usage{
pie(values, legendPosition = "bottom", width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{values}{A list of values with each it's own section of the pie chart.}

\item{legendPosition}{The position of the legend, default = "bottom".}

\item{width}{The width of the html widget.}

\item{height}{The width of the html widget.}

\item{elementId}{The element ID.}
}
\description{
An easy way to create c3 pie charts based on the rstudio js tutorials.
}
\examples{
pie(c(1,2,3,4))

pie(c(5,5,5))
}
