% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{pie.size}
\alias{pie.size}
\title{pie.size}
\usage{
pie.size(pier, outer = 90, inner = NULL, width = NULL, height = NULL)
}
\arguments{
\item{pier}{object}

\item{outer}{integer value converted to percentage size of pie radius within dimensions.
default set to 90 which is enough for most labels}

\item{inner}{integer value converted to percentage size of pie inner radius}

\item{width}{integer width, pixel-width}

\item{height}{integer height, pixel-height}
}
\examples{
\dontrun{
data.frame(value = c(40, 20, 30),
           label = c('red', 'green', 'blue'),
           color = c('red', 'green', 'blue')) \%>\%
  pier() \%>\%
  pie.size(outer = 80, inner = 50, width = 200, height = 200)
  }
}
