% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{pie.tooltips}
\alias{pie.tooltips}
\title{pie.tooltips}
\usage{
pie.tooltips(pier, enabled = TRUE, type = "placeholder",
  string = "{label}: {value}, {percentage}\%",
  placeholderParser = NULL, fadeInSpeed = 250,
  backgroundColor = "#000000", backgroundOpacity = 0.5,
  color = "#efefef", borderRadius = 2, font = "arial",
  fontSize = 10, padding = 4)
}
\arguments{
\item{pier}{object}

\item{enabled}{booleen; default = TRUE}

\item{type}{character type of tooltip used, must be 'placeholder' or 'caption'; default = 'placeholder'. For description see [d3pie docs](http://d3pie.org/#docs-tooltips)}

\item{string}{character gives methed to structure content and format of tootip; default = "\{label\}: \{value\}, \{percentage\}\%"}

\item{placeholderParser}{NOT YET FUNCTIONAL}

\item{fadeInSpeed}{integer value milliseconds; default = 250}

\item{backgroundColor}{character colour name or hex; default = '#000000',}

\item{backgroundOpacity}{numeric proportion must be between 0 and 1; default = 0.5,}

\item{color}{character colour name or hex; default = '#efefef',}

\item{borderRadius}{integer pixels; default = 2,}

\item{font}{character name; default = "arial",}

\item{fontSize}{integer pixels; default = 10,}

\item{padding}{integer pixels; default = 4}
}
\description{
Add tooltips to pie segments
}
\examples{
\dontrun{
data.frame(value = c(40, 20, 30),
           label = c('red', 'green', 'blue'),
           color = c('red', 'green', 'blue')) \%>\%
  pier() \%>\%
  pie.tooltips()
  }
}
