% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pier.R
\name{pier}
\alias{pier}
\title{pier}
\usage{
pier(data, width = NULL, height = NULL, header = NULL,
  sortOrder = "none", smallSegmentGrouping = FALSE, settings = NULL, ...)
}
\arguments{
\item{data}{data.frame must contain colnames value, label and color}

\item{width}{integer width, pixel-width}

\item{height}{integer height, pixel-height}

\item{header}{character, simple header above plot with no formatting. For more
control use `pier.title()`}

\item{sortOrder}{character must be one of 'none', 'random', 'value-asc', 'value-desc'
'label- asc', 'label-desc'}

\item{smallSegmentGrouping}{booleen option to combine segments below percent threshold}

\item{value}{smallSegmentGrouping option threshold for grouping segments}

\item{valueType}{smallSegmentGrouping option must be 'percentage' or 'value'}

\item{label}{character smallSegmentGrouping option default is 'Other'}

\item{color}{character smallSegmentGrouping option default is '#cccccc'}
}
\examples{
\dontrun{
data.frame(value = c(40, 20, 30),
           label = c('red', 'green', 'blue'),
           color = c('red', 'green', 'blue')) \%>\%
  pier()
  }
}
