% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{a \code{\link{d3pie}} object}

\item{rhs}{a pie settings function}
}
\description{
Imports the pipe operator from magrittr.
}
\examples{
\dontrun{
data.frame(value = c(40, 20, 30),
           label = c('red', 'green', 'blue'),
           color = c('red', 'green', 'blue')) \%>\%
  pier() \%>\%
  pie.size(inner = 50, outer = 90)
  }
}
